+++ crates/accelerate/src/unitary_synthesis.rs
@@ -234,8 +234,7 @@ fn apply_synth_sequence(
 ///     * 1q synthesis: OneQubitEulerDecomposer
 ///     * 2q synthesis: TwoQubitBasisDecomposer, TwoQubitControlledUDecomposer, XXDecomposer (Python, only if target is provided)
 ///     * 3q+ synthesis: QuantumShannonDecomposer (Python)
-/// This function is currently used in the Python `UnitarySynthesis`` transpiler pass as a replacement of the `run()`
-/// method of the `DefaultUnitarySynthesis` plugin.
+/// This function is currently used in the Python `UnitarySynthesis`` transpiler pass as a replacement for the `_run_main_loop` method.
 /// It returns a new `DAGCircuit` with the different synthesized gates.
 #[pyfunction]
 #[pyo3(name = "run_main_loop", signature=(dag, qubit_indices, min_qubits, target, basis_gates, coupling_edges, approximation_degree=None, natural_direction=None, pulse_optimize=None))]
